# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  description: This is Laravel Project APis using Sanctum
  version: 1.0.0
  title: Laravel Project API's
servers:
...
consumes:
  - application/json
  - application/xml
produces:
  - application/json
  - application/xml
  - application/pdf
...
schemes:
  - "http"
  - "https"
paths:
...
  /api/forgot-password:
    post:
      tags:
        - Auth
      summary: Send forgot password email
      operationId: forgotPassword
      requestBody:
        description: User email to receive link
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
        required: true
      responses:
        200:
          description: successful operation
        400:
          description: Invalid email supplied
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...